requirements.txt
Flask==2.1.2
waitress==2.1.1
setuptools==62.1.0
python-dotenv==0.20.0
Flask-BasicAuth==0.2.0